2. Парковка машин [#853 #214352]

+++ src/main/java/ood/lsp/Car.java
@@ -2,6 +2,6 @@ package ood.lsp;
 
 public interface Car {
 
-    double getSize();
+    int getSize();
 
 }
