+++ src/main/java/ood/lsp/Sedan.java
@@ -2,10 +2,10 @@ package ood.lsp;
 
 public class Sedan implements Car {
 
-    private final double size = 1;
+    private final int size = 1;
 
     @Override
-    public double getSize() {
+    public int getSize() {
         return size;
     }
 }
